examples/local/news-agent/requirements-ui.txt
chainlit
boto3
strands-agents
